+++ component/modal/ViewList.js
@@ -13,24 +13,19 @@ import {
   Provider,
 } from "react-native-paper";
 import { useNavigation } from "@react-navigation/native";
-import { TagPlusAPI } from "../../service/api";
+
 import { ScrollView } from "react-native";
 
 const ViewList = ({ data, handleDelete }) => {
   const [visible, setVisible] = React.useState(false);
-  const showModal = () => setVisible(true);
-  const hideModal = () => setVisible(false);
+  // const showModal = () => setVisible(true);
+  // const hideModal = () => setVisible(false);
   const navigation = useNavigation();
   const containerStyle = { backgroundColor: "white" };
 
-  const handleOrder = async () => {
-    const res = await TagPlusAPI(data);
-    navigation.navigate("Dashboard");
-    console.log(res, "res order");
-  };
   return (
     <ScrollView style={{ margin: 15 }}>
-      <List.Section style={{ maxHeight: 150 }}>
+      <List.Section>
         {data.map((list, index) => (
           <List.Item
             key={index}
@@ -44,28 +39,6 @@ const ViewList = ({ data, handleDelete }) => {
           />
         ))}
       </List.Section>
-
-      {data.length > 0 ? (
-        <Button
-          style={{
-            backgroundColor: "#6c65df",
-            padding: 8,
-            borderRadius: 15,
-            marginHorizontal: 10,
-            marginBottom: 10,
-          }}
-          onPress={handleOrder}
-          mode="contained"
-        >
-          Proceed to order
-        </Button>
-      ) : (
-        <Card style={{ margin: 15 }}>
-          <Text style={{ textAlign: "center", padding: 10 }}>
-            No Tags added to the List
-          </Text>
-        </Card>
-      )}
     </ScrollView>
   );
 };
